correciones de crear puesto v y listar, regionales crear y listar

+++ src/app/views/puestos-vig/list-puestos-vig/list-puestos-vig.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { PuestosVigilanciaService } from '../../../services/puestosvigilancia/puestosVig.service';
 import { MatTableDataSource } from '@angular/material/table';
 import { MatTableModule } from '@angular/material/table';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list-puestos-vig',
@@ -25,8 +26,9 @@ export class ListPuestosVigComponent implements OnInit {
   obtenerPuestos(): void {
     this.puestosService.obtenerPuestos().subscribe(
       (data) => {
-        this.puestos = data.data;
+        this.puestos = data.data[0];
         this.dataSource.data = this.puestos; // Actualiza el dataSource con los datos obtenidos
+        console.log('pv.', this.puestos)
       },
       (error) => {
         this.errorMessage = 'Error al obtener los puestos';
@@ -57,13 +59,36 @@ export class ListPuestosVigComponent implements OnInit {
   }
 
   eliminarPuesto(id: number): void {
-    this.puestosService.eliminarPuesto(id).subscribe(
-      (data) => {
-        this.obtenerPuestos();
-      },
-      (error) => {
-        this.errorMessage = 'Error al eliminar el puesto';
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: '¡No podrás revertir esto!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, eliminarlo'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.puestosService.eliminarPuesto(id).subscribe(
+          (data) => {
+            this.obtenerPuestos();
+            Swal.fire(
+              '¡Eliminado!',
+              'El puesto ha sido eliminado.',
+              'success'
+            );
+          },
+          (error) => {
+            this.errorMessage = 'Error al eliminar el puesto';
+            Swal.fire(
+              '¡Error!',
+              'Hubo un problema al eliminar el puesto.',
+              'error'
+            );
+          }
+        );
       }
-    );
+    });
   }
+  
 }
